+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   variable `partial-recall--log` is `t`.
 - Memories that have grown while remembering may shrink again after
   forgetting.
+- An overview can now be printed using `partial-recall-overview`.
 
 ### Changed